#2: Update and optimize lot of things
