+++ index.html
@@ -60,7 +60,7 @@
 
         <div id="left-content" class="both__content show" data-toggle-from="leftContent">
 
-          <div id="form-wrapper" class="form__wrapper">
+          <div id="form-wrapper" class="form__wrapper hidden">
             <!-- <label for="left-search-box">Cari buku yang belum dibaca</label> -->
 
             <form id="left-form" class="search__form">
@@ -78,6 +78,9 @@
 
             <div id="left-cards" class="cards">
               <!-- card content goes here -->
+              <p id="true-rakStillEmpty" style='color: #cbd5e1; margin: 16px auto 0;'>
+                Masih kosong nih, yuk isi... 😎😎😎
+              </p>
             </div>
           </div>
 
@@ -89,7 +92,7 @@
 
         <div id="right-content" class="both__content hidden" data-toggle-from="rightContent">
 
-          <div id="form-wrapper" class="form__wrapper">
+          <div id="form-wrapper" class="form__wrapper hidden">
             <!-- <label for="right-search-box">Cari buku yang sudah dibaca</label> -->
 
             <form id="right-form" class="search__form">
@@ -107,6 +110,9 @@
 
             <div id="right-cards" class="cards">
               <!-- card content goes here -->
+              <p id="false-rakStillEmpty" style='color: #cbd5e1; margin: 16px auto 0;'>
+                Masih kosong nih, yuk isi... 😎😎😎
+              </p>
             </div>
           </div>
